requirements.txt
coloredlogs>=5.2,<16.0
orderedattrdict>=1.4.3,<2.0
paho-mqtt>=1.2,<2.0
PyYAML>=6.0,<7.0
spidev>=3.5,<4.0
